+++ tests/integration/test_geometry_commands.py
@@ -61,3 +61,54 @@ def test_chamfer(modeler: Modeler):
     assert body2.volume.m == pytest.approx(
         Quantity(0.945333333333333333, UNITS.m**3).m, rel=1e-6, abs=1e-8
     )
+
+
+def test_fillet(modeler: Modeler):
+    """Test fillet on edge and face."""
+    design = modeler.create_design("chamfer")
+
+    body = design.extrude_sketch("box", Sketch().box(Point2D([0, 0]), 1, 1), 1)
+    assert len(body.faces) == 6
+    assert len(body.edges) == 12
+    assert body.volume.m == pytest.approx(Quantity(1, UNITS.m**3).m, rel=1e-6, abs=1e-8)
+
+    modeler.geometry_commands.fillet(body.edges[0], 0.1)
+    assert len(body.faces) == 7
+    assert len(body.edges) == 15
+    assert body.volume.m == pytest.approx(
+        Quantity(0.9978539816339744, UNITS.m**3).m, rel=1e-6, abs=1e-8
+    )
+
+    modeler.geometry_commands.fillet(body.faces[-1], 0.5)
+    assert len(body.faces) == 7
+    assert len(body.edges) == 15
+    assert body.volume.m == pytest.approx(
+        Quantity(0.946349540849362, UNITS.m**3).m, rel=1e-6, abs=1e-8
+    )
+
+    # multiple edges
+    body2 = design.extrude_sketch("box2", Sketch().box(Point2D([0, 0]), 1, 1), 1)
+    assert len(body2.faces) == 6
+    assert len(body2.edges) == 12
+    assert body2.volume.m == pytest.approx(Quantity(1, UNITS.m**3).m, rel=1e-6, abs=1e-8)
+
+    modeler.geometry_commands.fillet(body2.edges, 0.1)
+    assert len(body2.faces) == 26
+    assert len(body2.edges) == 48
+    assert body2.volume.m == pytest.approx(
+        Quantity(0.9755870138909422, UNITS.m**3).m, rel=1e-6, abs=1e-8
+    )
+
+    modeler.geometry_commands.fillet(body2.faces, 0.3)
+    assert len(body2.faces) == 26
+    assert len(body2.edges) == 48
+    assert body2.volume.m == pytest.approx(
+        Quantity(0.8043893421169303, UNITS.m**3).m, rel=1e-6, abs=1e-8
+    )
+
+    modeler.geometry_commands.fillet(body2.faces, 0.05)
+    assert len(body2.faces) == 26
+    assert len(body2.edges) == 48
+    assert body2.volume.m == pytest.approx(
+        Quantity(0.9937293491873294, UNITS.m**3).m, rel=1e-6, abs=1e-8
+    )
